move functions to Cpp and rework
