+++ fdrs_types.h
@@ -7,4 +7,11 @@ typedef struct __attribute__((packed)) DataReading {
     uint8_t type;
 } DataReading;
 
+
+typedef struct{
+    DataReading buffer[256];
+    uint16_t len = 0;
+    uint32_t time = 0;
+} DataReadingBuffer_t;
+
 #endif
